# api/swagger.yaml
...
swagger: '2.0'
...
info:
  version: "0.1.0"
  title: Patient Match Test Harness REST API
  description: |
    The vertical bar seems to denote a multi-line entry 
    that is "magically" terminated by the next field name.
    Note that __Markdown is supported__ in the text field.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  
# # # # # # # # # # # # # # # # # # # # # # # # # # # # #
#   Host, service Base Path, Schemes and Content Types
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # 
#host: api.instagram.com
schemes:
  - http
  - https
produces:
  - application/json
consumes:
  - application/json
  
# # # # # # # # # # # # # # # # # # # # # # # # # # # # #
#           Paths
# # # # # # # # # # # # # # # # # # # # # # # # # # # # # 
paths:
...
  /RecordMatchRun:
    get:
      operationId: getRecordMatchRuns
      summary: Get Record Match Runs
      tags:
        - RecordMatchRun
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/RecordMatchRun'
        400:
          description: Bad Request
        404:
          description: Not Found
    post:
      operationId: createRecordMatchRun
      summary: Create Record Match Run
      tags:
        - RecordMatchRun
      responses:
        201:
          description: Resource Created
        400:
          description: Bad Request
        500:
          description: Internal Server Error
  /RecordMatchRun/{id}:
    get:
      operationId: getRecordMatchRunById
      summary: Get one Record Match Run by Id
      tags:
        - RecordMatchRun
      parameters:
        - name: id
          in: path
          description: resource identifier 
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/RecordMatchRun'
    put:
      operationId: updateRecordMatchRun
      summary: Update Record Match Run
      tags:
        - RecordMatchRun
      parameters:
        - name: id
          in: path
          description: resource identifier 
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: '#/definitions/RecordMatchRun'
        400:
          description: Bad Request
        500:
          description: Internal Server Error
    delete:
      operationId: deleteRecordMatchRun
      summary: Delete Record Match Run
      tags:
        - RecordMatchRun
      parameters:
        - name: id
          in: path
          description: resource identifier 
          required: true
          type: string
      responses:
        204:
          description: No Content
        500:
          description: Internal Server Error
...
definitions:
...
  RecordMatchRequest:
    type: object
    description: |
      A Record Match Context encapsulates the FHIR message
      generated to request a record match operation.
    required:
    - id
    - message
    - submittedOn
    properties:
      id:
        type: string
      meta:
        $ref: '#/definitions/Meta'
      message:
        $ref: '#/definitions/Bundle'
      submittedOn:
        type: string
        format: date-time
        description: date and time at which the message was sent to the record match system (via a FHIR server acting as a message broker.
  RecordMatchResponse:
    type: object
    required:
    - id
    - message
    - receivedOn
    properties:
      id:
        type: string
      meta:
        $ref: '#/definitions/Meta'
      message:
        $ref: '#/definitions/Bundle'
      receivedOn:
        type: string
        format: date-time
        description: date and time at which the message was received from the record match system (via a FHIR server acting as a message broker.
  RecordMatchRun:
    type: object
    description: |
      A Record Match Run encapsulates all data associated
      with a record match run, including the context, the
      generated request message, response messages, metrics,
      and status.
    required:
    - id
    - matchingMode
    - masterRecordSetId
    properties:
      id:
        type: string
      matchingMode:
        type: string
        pattern: query|deduplication
        description: string literal, 'deduplication' or 'query'
      meta:
        $ref: '#/definitions/Meta'
      metrics:
        $ref: '#/definitions/RecordMatchRunMetrics'
      note:
        type: string
      recordMatchContextId:
        type: string
        enum: ['query', 'deduplication']
      recordMatchSystemInterfaceId:
        type: string
      masterRecordSetId:
        type: string
      queryRecordSetId:
        type: string
      recordResourceType:
        type: string
      request:
        $ref: '#/definitions/RecordMatchRequest'
      responses:
        type: array
        items:
          $ref: '#/definitions/RecordMatchResponse'
      status:
        type: array
        items:
          $ref: '#/definitions/RecordMatchRunStatusComponent'
  RecordMatchRunMetrics:
    type: object
    properties:
      f1:
        type: number
        format: float
        minimum: 0
        maximum: 1
      precision:
        type: number
        format: float
        minimum: 0
        maximum: 1
      recall:
        type: number
        format: float
        minimum: 0
        maximum: 1
      MAP:
        type: number
        format: float
      matchCount:
        type: integer
        minimum: 0
      truePositiveCount:
        type: integer
        minimum: 0
      falsePositiveCount:
        type: integer
        minimum: 0
...
  RecordMatchRunStatusComponent:
    type: object
...
  Bundle:
    type: object
  Meta:
    type: object
    properties:
      createdOn: 
        description: date and time at which the resource was created.
        type: string
        format: date-time
      lastUpdatedOn: 
        type: string
        format: date-time
...